WPF Application/Converters/NullableToBoolean.cs
using System;
using System.Collections.Generic;
using System.Globalization;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using System.Windows;

namespace Right_Click_Commands.WPF.Converters
{
    public class NullableToBoolean : ConverterBase<NullableToBoolean>
    {
        //  Methods
        //  =======

        public override object Convert(object value, Type targetType, object parameter, CultureInfo culture)
        {
            return value != null;
        }

        public override object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
        {
            if (!(value is bool))
                throw new ArgumentException("The given value must be a boolean");

            return ((bool)value) ? new object() : null;
        }
    }
}
